+++ packages/datasource-mongoose/test/integration/many-to-many/collection_delete.test.ts
@@ -4,8 +4,8 @@ import { Filter, Projection } from '@forestadmin/datasource-toolkit';
 import * as factories from '@forestadmin/datasource-toolkit/dist/test/__factories__';
 import { Connection, Types } from 'mongoose';
 
+import setupWith2ManyToManyRelations from './_build-models';
 import MongooseDatasource from '../../../src/datasource';
-import { setupWith2ManyToManyRelations } from '../_build-models';
 
 describe('MongooseCollection', () => {
   let connection: Connection;
