+++ packages/datasource-mongoose/test/integration/review/collection_create.test.ts
@@ -4,8 +4,8 @@ import { ValidationError } from '@forestadmin/datasource-toolkit';
 import * as factories from '@forestadmin/datasource-toolkit/dist/test/__factories__';
 import { Connection } from 'mongoose';
 
+import setupReview from './_build-models';
 import MongooseDatasource from '../../../src/datasource';
-import { setupReview } from '../_build-models';
 
 const caller = factories.caller.build();
