+++ packages/datasource-mongoose/test/integration/review/collection_list.test.ts
@@ -11,8 +11,8 @@ import {
 import * as factories from '@forestadmin/datasource-toolkit/dist/test/__factories__';
 import { Connection, Types } from 'mongoose';
 
+import setupReview from './_build-models';
 import MongooseDatasource from '../../../src/datasource';
-import { setupReview } from '../_build-models';
 
 describe('MongooseCollection', () => {
   let connection: Connection;
