+++ packages/datasource-mongoose/test/integration/review/collection_update.test.ts
@@ -4,8 +4,8 @@ import { Filter, Projection } from '@forestadmin/datasource-toolkit';
 import * as factories from '@forestadmin/datasource-toolkit/dist/test/__factories__';
 import { Connection, Schema, Types } from 'mongoose';
 
+import setupReview from './_build-models';
 import MongooseDatasource from '../../../src/datasource';
-import { setupReview } from '../_build-models';
 
 describe('MongooseCollection > update', () => {
   let connection: Connection;
